+++ example/ios/Runner/AppDelegate.m
@@ -1,5 +1,6 @@
 #import "AppDelegate.h"
 #import "GeneratedPluginRegistrant.h"
+#import <DOT/DOT.h>
 
 @implementation AppDelegate
 
@@ -7,6 +8,18 @@
     didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
   [GeneratedPluginRegistrant registerWithRegistry:self];
   // Override point for customization after application launch.
+  
+
+  // S: Wisetracker SDK init
+  [DOT initialization:launchOptions application:application];
+  #ifdef DEBUG
+    [DOT checkDebugMode:true];
+  #else
+    [DOT checkDebugMode:false];
+  #endif
+  // E: Wisetracker SDK init
+
+
   return [super application:application didFinishLaunchingWithOptions:launchOptions];
 }