+++ CHANGELOG.md
@@ -16,4 +16,4 @@
 
 ## 1.0.4 
 
-* Push Message 관련 api 추가
+* Push Message 관련 api 추가, example 코드 업데이트